refactor: Refactor the test file to ts, use jest instead of tape, modify tsconfig to support default import
