+++ package.json
@@ -30,16 +30,17 @@
     "index.ts"
   ],
   "devDependencies": {
-    "@types/tape": "^4.0.0",
+    "@types/jest": "^26.0.23",
     "c8": "^7.0.0",
     "chalk": "^4.0.0",
+    "jest": "^26.6.3",
     "nyc": "^15.0.0",
     "prettier": "^2.0.0",
     "remark-cli": "^9.0.0",
     "remark-preset-wooorm": "^8.0.0",
     "rimraf": "^3.0.0",
     "strip-ansi": "^6.0.0",
-    "tape": "^5.0.0",
+    "ts-jest": "^26.5.6",
     "type-coverage": "^2.0.0",
     "typescript": "^4.0.0",
     "xo": "^0.38.0"
@@ -48,9 +49,12 @@
     "prepack": "npm run build && npm run format",
     "build": "rimraf \"*.d.ts\" && tsc && type-coverage",
     "format": "remark . -qfo && prettier . -w --loglevel warn && xo --fix",
-    "test-api": "node test.js",
-    "test-coverage": "c8 --check-coverage --branches 100 --functions 100 --lines 100 --statements 100 --reporter lcov node test.js",
-    "test": "npm run build && npm run format && npm run test-coverage"
+    "test-api": "jest --all",
+    "test-coverage": "jest --all --collect-coverage",
+    "test": "npm run format && npm run test-coverage"
+  },
+  "jest": {
+    "preset": "ts-jest"
   },
   "remarkConfig": {
     "plugins": [
